docs(PPDSC-2191): merge and best practice files moved to forms folder
