+++ site/components/illustrations/patterns/forms/best-practice/labels/optional.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import {Svg} from '../../../svg';
-import {Path} from '../../../path';
+import {Svg} from '../../../../svg';
+import {Path} from '../../../../path';
 
 export const Optional: React.FC = () => (
   <Svg
